Try adding file import.
Change the code to import the data.txt file.
Add Bias Term to Method
Use Sigmoid Activation Function in Method
Change num_epochs=1 and alpha=0.001 in the call
Use percent_error calculation. By checking what percentage of the number of incorrect predictions is out of the total amount of data.